Resolved issue #2978 Refactor Commander
