+++ commander/pom.xml
@@ -39,70 +39,33 @@
       <artifactId>junit-jupiter-engine</artifactId>
       <scope>test</scope>
     </dependency>
+      <dependency>
+          <groupId>org.slf4j</groupId>
+          <artifactId>slf4j-api</artifactId>
+          <version>${slf4j.version}</version>
+      </dependency>
   </dependencies>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>Employee</id>
-            <configuration>
-              <archive>
-                <manifest>
-                  <mainClass>com.iluwatar.commander.AppEmployeeDbFailCases</mainClass>
-                </manifest>
-              </archive>
-              <finalName>${project.artifactId}-EmployeeDBFailCase</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>Message</id>
-            <configuration>
-              <archive>
-                <manifest>
-                  <mainClass>com.iluwatar.commander.AppMessagingFailCases</mainClass>
-                </manifest>
-              </archive>
-              <finalName>${project.artifactId}-MessagingFailCase</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>Payment</id>
-            <configuration>
-              <archive>
-                <manifest>
-                  <mainClass>com.iluwatar.commander.AppPaymentFailCases</mainClass>
-                </manifest>
-              </archive>
-              <finalName>${project.artifactId}-PaymentFailCase</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>Queue</id>
-            <configuration>
-              <archive>
-                <manifest>
-                  <mainClass>com.iluwatar.commander.AppQueueFailCases</mainClass>
-                </manifest>
-              </archive>
-              <finalName>${project.artifactId}-QueueFailCase</finalName>
-            </configuration>
-          </execution>
-          <execution>
-            <id>Shipping</id>
-            <configuration>
-              <archive>
-                <manifest>
-                  <mainClass>com.iluwatar.commander.AppShippingFailCases</mainClass>
-                </manifest>
-              </archive>
-              <finalName>${project.artifactId}-ShippingFailCase</finalName>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
+                <executions>
+                    <execution>
+                        <id>allCases</id>
+                        <configuration>
+                            <archive>
+                                <manifest>
+                                    <mainClass>com.iluwatar.commander.AppAllCases</mainClass>
+                                </manifest>
+                            </archive>
+                            <finalName>${project.artifactId}-AllCases</finalName>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
 </project>
